aperantly solved SSR error

+++ pictext_proyect/src/app/home/home.component.ts
@@ -1,6 +1,7 @@
-import { Component, inject } from '@angular/core';
+import { Component, Inject, inject, PLATFORM_ID } from '@angular/core';
 import { RouterLink, Router } from '@angular/router';
 import { UserService } from '../user.service';
+import { isPlatformBrowser } from '@angular/common';
 
 @Component({
   selector: 'app-home',
@@ -12,14 +13,22 @@ import { UserService } from '../user.service';
 export class HomeComponent {
 	userService = inject(UserService)
 
-	constructor(private router: Router) { }
+	constructor(
+		private router: Router,
+		@Inject(PLATFORM_ID) private platformId: Object	
+	) { }
 
 	getData() {
-    return sessionStorage.getItem('name');
+    if (isPlatformBrowser(this.platformId)) {
+      return sessionStorage.getItem('name');
+    }
+		return null;
   }
 
 	deleteData() {
-    sessionStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      sessionStorage.clear();
+    }
   }
 
 	handleLogin(){
